Eliminate 421 MinGW build warnings (98.6% reduction)
Mechanical fixes across 62 files to clear C++20-deprecation, NULL-misuse, and
miscellaneous warnings flagged by the i686-w64-mingw32 GCC 13 build:

- _struct.h, WSclient.h: drop trailing tokens after #endif; switch memset
  fill values from NULL to 0 (these two headers alone accounted for ~1500
  warnings before).
- 158 NULL-as-int sites converted to 0 (or L'\0' for wchar_t) at compiler-
  flagged lines only.
- 218 deprecated enum+enum arithmetic sites wrapped with static_cast<int>(...)
  on the unnamed-enum operand (MODEL_TYPE_CHARM_MIXWING + EWS_*,
  MODEL_BODY_* + c->SkinIndex, MODEL_X - MODEL_ITEM, etc).
- 9 trailing #endif tokens in GameShop/MsgBoxIGS*.cpp.
- 3 narrowing conversions in POINT/SIZE/RECT brace-initializers.
- 3 enum-compare sites cast to int.
- 6 BYTE-typed switches with negative case literals (-1, -2) → 255, 254.
- 5 delete-of-void* sites cast to the concrete pointer type before delete
  (xstreambuf BYTE*, CSParts CSIPartsMDL*).
- 1 VLA-style `new (T[n])` → `new T[n]` (UIMng.cpp).
- Stray `]` after #include in ZzzEffect.cpp.
- Redundant `extern` on a definition in UIGuildInfo.cpp.
- `#pragma once` removed from main-file .cpp's (NewUIGroup, NewUITextBox).
- stdafx.h: NOMINMAX guarded with #ifndef so libstdc++ pre-define no longer
  produces a redefine warning.

Six warnings remain — all flag real bugs that need human review rather than
mechanical changes:
  TextureScript.cpp:58,63    case 'DC'/'DT' on a single-char switch (dead code)
  ZzzInterface.cpp:3883      'and' of mutually exclusive equal-tests is always 0
  UIWindows.cpp:4215         delete on pointer to a range-for copy member (UB)

Verified locally with the same toolchain CI uses (i686-w64-mingw32, Ninja,
Release, BUILD_TESTING=OFF). Build links cleanly; final warning count: 6.

+++ src/source/CharMakeWin.cpp
@@ -461,7 +461,7 @@ void CCharMakeWin::RenderControls()
 void CCharMakeWin::SelectCreateCharacter()
 {
     CharacterView.Class = m_nSelJob;
-    CreateCharacterPointer(&CharacterView, MODEL_FACE + CharacterView.Class, 0, 0);
+    CreateCharacterPointer(&CharacterView, static_cast<int>(MODEL_FACE) + CharacterView.Class, 0, 0);
     CharacterView.Object.Kind = 0;
     SetAction(&CharacterView.Object, 1);
 }
